fix: Mouse responds only to left,middle, right press

+++ input.h
@@ -7,9 +7,9 @@ void HandleInput(void);
 bool KeyWentDown(SDL_Keycode keyCode);
 bool KeyWentUp(SDL_Keycode keyCode);
 bool KeyPressed(SDL_Keycode keyCode);
-bool MouseWentDown(Uint8 button);
-bool MouseWentUp(Uint8 button);
-bool MousePressed(Uint8 button);
+bool MouseWentDown(int button);
+bool MouseWentUp(int button);
+bool MousePressed(int button);
 bool RightMouseWentDown(void);
 bool ShouldClose(void);
